requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
firebase-admin==6.5.0
razorpay==1.4.2
twilio==9.4.1
loguru==0.7.3
pyyaml==6.0.2
cryptography==44.0.0
